plyFind/python_backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
openai>=1.3.0
deep-translator>=1.11.4
gtts>=2.3.2
SoundFile>=0.12.1
sounddevice>=0.4.6
numpy>=1.24.3
python-multipart>=0.0.6